+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+
+## 2022-10-27
+
+---
+
+### Added:
+* Ability for force quit if shutdown going for more than 30 seconds
+
 ## 2022-10-26
 
 ---